[RPI] secu (data) : use INJECT_ALL or INJECT_MINIMUM_ACCESS_RIGHT in production to inject data (no default)

+++ manifmaker/server/main.js
@@ -12,9 +12,29 @@ Meteor.startup(function () {
 
     ServerService.addCollectionHooks();
 
-    InjectDataServerService.deleteAll();
-    InjectDataServerService.initAccessRightData();
-    InjectDataServerService.injectAllData();
+    var injectAll = process.env.INJECT_ALL;
+    if (typeof(injectAll) !== 'undefined') {
+        console.info("Meteor.startup : injectAll trigger by ENV (deleteAll, initAccessRightData, injectAllData)");
+        InjectDataServerService.deleteAll();
+        InjectDataServerService.initAccessRightData();
+        InjectDataServerService.injectAllData();
+    }
+    var initAccessRight = process.env.INJECT_MINIMUM_ACCESS_RIGHT;
+    if (typeof(initAccessRight) !== 'undefined') {
+        console.info("Meteor.startup : initAccessRight trigger by ENV (deleteAll, initAccessRightData)");
+        InjectDataServerService.deleteAll();
+        InjectDataServerService.initAccessRightData();
+    }
+
+    if(Meteor.isDevelopment){
+        //specific to the dev needs
+        console.info("Meteor.startup : isDevelopment, injecting or not");
+        InjectDataServerService.deleteAll();
+        InjectDataServerService.initAccessRightData();
+        InjectDataServerService.injectAllData();
+    }
+
+
 
     Meteor.isStartingUp = false;
 });
